+++ luci-app-rakitanmanager/root/usr/share/rakitanmanager/modem-hp.sh
@@ -13,6 +13,13 @@ if ! command -v adb &> /dev/null; then
     exit 1
 fi
 
+# Periksa koneksi ADB
+devices=$(adb devices | sed -n '2p')
+if [[ -z "$devices" || "$devices" == *"unauthorized"* ]]; then
+    echo "ADB device not connected or unauthorized."
+    exit 1
+fi
+
 if [ -z "$1" ]; then
 	log "ADBID tidak disetel, menggunakan default..."
 	ADBID=$(adb devices | grep 'device' | grep -v 'List of' | awk {'print $1'}) # Default device_id if $1 unset
@@ -25,23 +32,12 @@ if [ "$2" = "restart" ]; then
     do
         log "Menghubungkan ke perangkat ${IPX}..." 
         log "Mode pesawat akan diaktifkan dalam 3 detik..."
-        if [[ "$(adb -s ${IPX} shell settings get global airplane_mode_on)" == "0" ]]; then
-            adb -s "$IPX" settings put global airplane_mode_on 1 &>/dev/null
-            adb -s "$IPX" am broadcast -a android.intent.action.AIRPLANE_MODE --ez state true &>/dev/null
-        fi
-        if [[ "$(adb -s ${IPX} shell cmd connectivity airplane-mode)" == "disabled" ]]; then
-            adb -s "$IPX" shell cmd connectivity airplane-mode enable &>/dev/null
-        fi
+        adb -s "$IPX" shell settings put global airplane_mode_on 1 &>/dev/null
+        adb -s "$IPX" shell am broadcast -a android.intent.action.AIRPLANE_MODE --ez state true &>/dev/null
         sleep "3" &>/dev/null
-
         log "Menonaktifkan mode pesawat untuk mendapatkan IP baru dan jaringan yang diperbarui..."
-        if [[  "$(adb -s ${IPX} shell settings get global airplane_mode_on)" == "1" ]]; then
-            adb -s "$IPX" settings put global airplane_mode_on 0 &>/dev/null
-            adb -s "$IPX" am broadcast -a android.intent.action.AIRPLANE_MODE --ez state false &>/dev/null
-        fi
-        if [[ "$(adb -s ${IPX} shell cmd connectivity airplane-mode)" == "enabled" ]]; then
-            adb -s "$IPX" shell cmd connectivity airplane-mode disable &>/dev/null
-        fi
+        adb -s "$IPX" shell settings put global airplane_mode_on 0 &>/dev/null
+        adb -s "$IPX" shell am broadcast -a android.intent.action.AIRPLANE_MODE --ez state false &>/dev/null
         log "ID [${IPX}] : Penyegaran jaringan selesai...!!"
         exit 0
     done
@@ -51,20 +47,14 @@ if [ "$2" = "myip" ]; then
     # Jalankan perintah adb untuk mendapatkan informasi jaringan perangkat
     for IPX in ${ADBID}
     do
-        network_info=$(adb -s "${IPX} shell ip addr show")
-
-        # Cari baris yang berisi alamat IP
-        ip_address_line=$(echo "$network_info" | grep 'inet ' | grep -v '127.0.0.1')
-
-        # Jika ditemukan baris yang berisi alamat IP
-        if [ -n "$ip_address_line" ]; then
-            # Ambil alamat IP dari baris tersebut
-            ip_address=$(echo "$ip_address_line" | awk '{print $2}')
-            echo "$ip_address"
-            exit 0
-        else
-            log "Tidak dapat menemukan alamat IP perangkat."
-            exit 1
+        ip_output=$(adb -s "$IPX" shell ip addr show)
+        ip_addr=$(echo "$ip_output" | grep -Eo 'inet (addr:)?([0-9]*\.){3}[0-9]*' | grep -v '127.0.0.1' | awk '{print $2}' | head -n 1)
+        
+        # Jika alamat IP kosong, tetapkan "Tidak tersedia"
+        if [[ -z "$ip_addr" ]]; then
+            ip_addr="Unavailable"
         fi
+        
+        log "New IP: $ip_addr"
     done
 fi